requirements.txt
# ServisHub - Production Dependencies
# Python 3.11+

# Web Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Database
SQLAlchemy==2.0.23
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
alembic==1.13.0

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2

# Validation
pydantic==2.5.2
email-validator==2.1.0

# Cache & Queue
redis==5.0.1
celery==5.3.4

# File Upload
cloudinary==1.37.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
python-slugify==8.0.1

# API
flask-cors==4.0.0

# Logging
structlog==23.2.0
